Extract configuration-specific functionality into "configs.py" module.
