+++ aanalytics2/__init__.py
@@ -1,2 +1,4 @@
 __version__ = "0.1.2"
+
 from .aanalytics2 import *
+from .configs import *
